backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
aiohttp
websockets
pyyaml
# awsiot e awscrt - da aggiungere in seguito per supporto MQTT AWS
